+++ CHANGELOG.rst
@@ -1,12 +1,25 @@
 Releases
 ========
 
-dev
----
+v0.7.0
+------
+Features
+++++++++
+* transfer function class
+* parallelization for transfer function
+* numerical transfer function calculation from cross and auto mock
+* parameter fitting sampler class
+* parallelization for parameter fitting
+* support emcee and nautilus for sampling
+
 Enhancements
 ++++++++++++
 * allow k-mode cut in 1D power spectrum
 * add validation tests 00 and 01 in paper
+
+Bugfixes
+++++++++
+* fix a bug in FPS initialization
 
 v0.6.1
 ------